Reorganized some of the proofs to improve abound.
I modified the weight minimization proof to make it apply to
pseudo-generators and state its bound in terms of the number
of pseudo-generators.  I then proved a few Lemmas that let
me state exactly how many there are, and am about to use this
in a Proposition that puts a stricter bound on how much time
it needs to search for the minimal weight undetectable error.